+++ api-server/source/main.ts
@@ -8,6 +8,7 @@ import yargs from "yargs";
 import { dumpConfiguration } from "./config.ts";
 import { runMigrations, useAppConfig, useTokens } from "./lib/mod.ts";
 import { runServer } from "./server.ts";
+import { notifyCommand } from "./notifications/notify-command.ts";
 
 const cli = yargs(process.argv.slice(2))
   .help()
@@ -48,6 +49,18 @@ cli.command(
     if (useAppConfig().env !== "development") console.log("not_allowed");
     else console.log(await useTokens().sign("admin", { userId: args.user_id }));
   },
+);
+
+cli.command(
+  "notify",
+  "start a process that periodically sends notifications in the queue",
+  (yargs) =>
+    yargs
+      .option("interval", { type: "number", default: 30_000 })
+      .option("grace", { type: "number", default: 1_000 })
+      .option("dry-run", { type: "boolean", default: false })
+      .option("forever", { type: "boolean", default: false }),
+  (args) => notifyCommand(args),
 );
 
 try {